requirements.txt
streamlit>=1.32
tmdbv3api>=1.8.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
pandas>=2.0.0